Message Queue changes (1) Limit the size of the message queue (2) Limit the number of message queues open simultaneously
What I found is that after 10 message queues were opened, I would start seeing errors. There
is a default system wide limit of 256 and a per-process limit of 10, plus a limit of 819200
total (aggregate) size.  I started with the logic to close message queues when the list got
too big but that didn't do it.  I decreased the size of them as well.  The inital pass of that
did not work either, but that is because I had removed the unlink logic.

Needless to say, I'm not especially impressed with the POSIX mq implementation.  I've made it
work for now, but this suggests I will need to revisit this again.  I could use a hybrid
where I connect and request via the mq, but get the response via shared memory.